fix check

+++ app/src/main/java/com/tqd/duy/models/Food.java
@@ -1,9 +1,6 @@
 package com.tqd.duy.models;
 
 import android.app.Activity;
-import android.os.Parcel;
-import android.os.Parcelable;
-
 import com.tqd.duy.coffeemanagement.R;
 import java.io.Serializable;
 import java.util.ArrayList;
@@ -14,7 +11,7 @@ import java.util.ArrayList;
 
 public class Food implements Serializable {
     private String nameFood;
-    private Double priceFood;
+    private int priceFood;
     private int numberFood;
 
     @Override
@@ -31,11 +28,11 @@ public class Food implements Serializable {
         this.nameFood = nameFood;
     }
 
-    public Double getPriceFood() {
+    public int getPriceFood() {
         return priceFood;
     }
 
-    public void setPriceFood(Double priceFood) {
+    public void setPriceFood(int priceFood) {
         this.priceFood = priceFood;
     }
 
@@ -48,7 +45,7 @@ public class Food implements Serializable {
     }
 
 
-    public Food( String nameFood, Double priceFood, int numberFood) {
+    public Food( String nameFood, int priceFood, int numberFood) {
         this.nameFood = nameFood;
         this.priceFood = priceFood;
         this.numberFood = numberFood;
@@ -59,16 +56,16 @@ public class Food implements Serializable {
 
     public ArrayList<Food> getSampleFood(Activity context) {
         ArrayList<Food> listFood = new ArrayList<>();
-        listFood.add(new Food(context.getResources().getString(R.string.Black_coffee), 10000.0,0));
-        listFood.add(new Food(context.getResources().getString(R.string.Black_coffee_SG), 12000.0,0));
-        listFood.add(new Food(context.getResources().getString(R.string.Coffee_milk), 15000.0,0));
-        listFood.add(new Food(context.getResources().getString(R.string.Coffee_milk_SG), 17000.0,0));
-        listFood.add(new Food(context.getResources().getString(R.string.Orange_juice), 15000.0,0));
-        listFood.add(new Food(context.getResources().getString(R.string.Lemon_juice), 12000.0,0));
-        listFood.add(new Food(context.getResources().getString(R.string.Pepsi), 8000.0,0));
-        listFood.add(new Food(context.getResources().getString(R.string.Yellow_sting), 9000.0,0));
-        listFood.add(new Food(context.getResources().getString(R.string.Strawberry_sting), 9000.0,0));
-        listFood.add(new Food(context.getResources().getString(R.string.Aquafina), 5000.0,0));
+        listFood.add(new Food(context.getResources().getString(R.string.Black_coffee), 10000,0));
+        listFood.add(new Food(context.getResources().getString(R.string.Black_coffee_SG), 12000,0));
+        listFood.add(new Food(context.getResources().getString(R.string.Coffee_milk), 15000,0));
+        listFood.add(new Food(context.getResources().getString(R.string.Coffee_milk_SG), 17000,0));
+        listFood.add(new Food(context.getResources().getString(R.string.Orange_juice), 15000,0));
+        listFood.add(new Food(context.getResources().getString(R.string.Lemon_juice), 12000,0));
+        listFood.add(new Food(context.getResources().getString(R.string.Pepsi), 8000,0));
+        listFood.add(new Food(context.getResources().getString(R.string.Yellow_sting), 9000,0));
+        listFood.add(new Food(context.getResources().getString(R.string.Strawberry_sting), 9000,0));
+        listFood.add(new Food(context.getResources().getString(R.string.Aquafina), 5000,0));
         return listFood;
     }
